refactor(basic): ui 컴포넌트와 상태 로직 분리
- 상태 값과 함수들을 분리하여 상위 컴포넌트에서 관리하도록 수정
- UI 관련 컴포넌트는 props 기반으로 재구성
